requirements.txt
discord.py
discord.py[voice]
google-genai
feedparser
requests
python-dotenv
pytubefix
pytz
yt-dlp